chore: Fix a bunch of compile warnings (#2748)
* chore: fix warnings for simu and firmware, part 1

* chore: fix warnings

+++ radio/src/gui/colorlcd/model_mixes.cpp
@@ -267,7 +267,7 @@ void MixLineButton::refresh()
 
   if (line.name[0]) {
     int cnt = lv_snprintf(s, maxlen, "%.*s ", (int)sizeof(line.name), line.name);
-    if (cnt >= maxlen) maxlen = 0;
+    if ((size_t)cnt >= maxlen) maxlen = 0;
     else { maxlen -= cnt; s += cnt; }
   }
 
@@ -275,13 +275,13 @@ void MixLineButton::refresh()
     if (line.swtch) {
       char* sw_pos = getSwitchPositionName(line.swtch);
       int cnt = lv_snprintf(s, maxlen, "%s ", sw_pos);
-      if (cnt >= maxlen) maxlen = 0;
+      if ((size_t)cnt >= maxlen) maxlen = 0;
       else { maxlen -= cnt; s += cnt; }
     }
     if (line.curve.value != 0) {
       getCurveRefString(s, maxlen, line.curve);
       int cnt = strnlen(s, maxlen);
-      if (cnt >= maxlen) maxlen = 0;
+      if ((size_t)cnt >= maxlen) maxlen = 0;
       else { maxlen -= cnt; s += cnt; }
     }
   }
